+++ plugins/green_vocoder/src/ui/channel_vocoder.cpp
@@ -94,7 +94,7 @@ void ChannelVocoder::paint(juce::Graphics& g) {
     float x = bb.getX();
     for (int i = 0; i < nbands; ++i) {
         juce::Rectangle<float> rect{x + width * 0.25f, bb.getY(), width * 0.5f, bb.getHeight()};
-        float gain = vocoder_.GetBinPeak(static_cast<size_t>(i))[0];
+        float gain = vocoder_.GetBinPeak(i)[0];
 
         float db_gain = 20.0f * std::log10(gain + 1e-10f);
         db_gain = std::clamp(db_gain, down, up);
